requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
boto3
chromadb
pydantic
# transformers
# torch
mysql-connector-python
# sentence_transformers
pandas
openai
